+++ Worms_SFML/Worms_SFML/physics.hpp
@@ -2,9 +2,9 @@
 #ifndef PHYSICS_H
 #define PHYSICS_H
 //ѕеревод из метрической системы Box2D в пиксели
-extern const int SCALE;
+extern const float SCALE;
 //ѕеревод из радиан в градусы
-extern const int RAD;
+extern const double RAD;
 //¬спомогательный массив дл€ хранени€ физики земли
 extern bool groundSavePhysics[2500][1000];
 //’ранение тел каждого черв€
